backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
cffi==1.17.1
cryptography==42.0.5
Django==5.0.3
django-browser-reload==1.18.0
django-cors-headers==4.3.1
django-csp==4.0
django-tailwind==4.2.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
wheel==0.45.1
